requirements.txt
huggingface_hub
transformers>=4.56.0
pyannote-audio==3.3.2
pydub
tqdm
qwen-asr
